change files structure and import statemets
