+++ backend/hct_mis_api/api/endpoints/__init__.py
@@ -1,4 +1,3 @@
 from hct_mis_api.api.endpoints.core import *  # noqa: F401, F403
 from hct_mis_api.api.endpoints.lookups import *  # noqa: F401, F403
-from hct_mis_api.api.endpoints.program import *  # noqa: F401, F403
 from hct_mis_api.api.endpoints.rdi import *  # noqa: F401, F403
