+++ backend/hct_mis_api/api/endpoints/core/__init__.py
@@ -1,3 +1,1 @@
-from hct_mis_api.api.endpoints.core.business_area import (  # noqa: F401
-    BusinessAreaListView,
-)
+from hct_mis_api.api.endpoints.core.views import BusinessAreaListView  # noqa: F401
