+++ backend/hct_mis_api/api/endpoints/core/serializers.py
@@ -1,7 +1,5 @@
 from rest_framework import serializers
-from rest_framework.generics import ListAPIView
 
-from hct_mis_api.api.endpoints.base import HOPEAPIView
 from hct_mis_api.apps.core.models import BusinessArea
 
 
@@ -18,8 +16,3 @@ class BusinessAreaSerializer(serializers.ModelSerializer):
             "is_split",
             "active",
         )
-
-
-class BusinessAreaListView(HOPEAPIView, ListAPIView):
-    serializer_class = BusinessAreaSerializer
-    queryset = BusinessArea.objects.all()
