--- backend/hct_mis_api/api/endpoints/program/__init__.py
@@ -1,3 +1,0 @@
-from hct_mis_api.api.endpoints.program.program import (  # noqa: F401
-    ProgramGlobalListView,
-)
