+++ backend/hct_mis_api/api/endpoints/program/serializers.py
@@ -1,7 +1,5 @@
 from rest_framework import serializers
-from rest_framework.generics import ListAPIView
 
-from hct_mis_api.api.endpoints.base import HOPEAPIView
 from hct_mis_api.apps.program.models import Program
 
 
@@ -25,8 +23,3 @@ class ProgramGlobalSerializer(serializers.ModelSerializer):
             "population_goal",
             "business_area_code",
         )
-
-
-class ProgramGlobalListView(HOPEAPIView, ListAPIView):
-    serializer_class = ProgramGlobalSerializer
-    queryset = Program.objects.all()
